backend/pyproject.toml
[project]
name = "camel-community-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.3",
    "celery[redis]>=5.4",
    "django-redis>=5.4",
    "psycopg[binary]>=3.2",
    "django-allauth[mfa]>=65.0",
    "djangorestframework-simplejwt>=5.3",
    "django-storages[s3]>=1.14",
    "django-cors-headers>=4.4",
    "django-ratelimit>=4.1",
    "django-extensions>=3.2",
    "meilisearch>=0.31",
    "stripe>=10.0",
    "sentry-sdk[django]>=2.0",
    "Pillow>=11.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest-django",
    "ruff",
    "mypy",
    "django-stubs",
    "factory-boy",
]

[tool.ruff]
line-length = 88

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
